+++ src/ui/common/viewers/Basic.h
@@ -47,6 +47,7 @@ void Viewer(const char* label, void* const& obj);
 void Viewer(const char* label, const Eigen::Quaternionf& quat);
 void Viewer(const char* label, const hh::fnd::WorldPosition& worldPos);
 void Viewer(const char* label, const csl::math::Transform& transform);
+void Viewer(const char* label, hh::game::GameObject* const& gameObject);
 void Viewer(const char* label, const csl::ut::Color8& color);
 void Viewer(const char* label, const csl::ut::Color<float>& color);