+++ src/ui/game-services/game-service-inspectors/PathManager.cpp
@@ -1,0 +1,13 @@
+#include "PathManager.h"
+#include <ui/common/viewers/Basic.h>
+
+void RenderGameServiceInspector(hh::path::PathManager& service) {
+	for (auto* pathComponent : service.pathComponents) {
+		if (ImGui::TreeNode(pathComponent, "%s", pathComponent->GetName())) {
+			Viewer("Name", pathComponent->GetName());
+			Viewer("Owner object", pathComponent->GetOwnerGameObject());
+
+			ImGui::TreePop();
+		}
+	}
+}