+++ src/ui/game-services/game-service-inspectors/PathManager.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void RenderGameServiceInspector(hh::path::PathManager& service);