+++ src/ui/operation-modes/modes/object-inspection/component-inspectors/PathComponent.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void RenderComponentInspector(hh::path::PathComponent& component);